section-8--database-with-python--sqlite3-pymysql/requirements.txt
PyMySQL[rsa]==1.1.2
python-dotenv==1.1.1
